+++ Tests/Unit/MSFT_WaitForADDomain.Tests.ps1
@@ -168,6 +168,73 @@ try
                     }
                 }
             }
+
+            Context 'When the Find-DomainController throws an authentication exception' {
+                BeforeAll {
+                    Mock -CommandName Find-DomainController -MockWith {
+                        $exceptionWithInnerException = New-Object -TypeName 'System.Management.Automation.MethodInvocationException' `
+                        -ArgumentList @(
+                            'The user name or password is incorrect.',
+                            (New-Object -TypeName 'System.Security.Authentication.AuthenticationException')
+                        )
+
+                        $newObjectParameters = @{
+                            TypeName     = 'System.Management.Automation.ErrorRecord'
+                            ArgumentList = @(
+                                $exceptionWithInnerException,
+                                'AuthenticationException',
+                                'InvalidOperation',
+                                $null
+                            )
+                        }
+
+                        throw New-Object @newObjectParameters
+                    }
+                }
+
+                BeforeEach {
+                    $getTargetResourceParameters = $mockDefaultParameters.Clone()
+                    $getTargetResourceParameters['Credential'] = $mockDomainUserCredential
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is not specified' {
+                    It 'Should throw the correct error' {
+                        { Get-TargetResource @getTargetResourceParameters } | Should -Throw 'The user name or password is incorrect.'
+                    }
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is set to $false' {
+                    BeforeEach {
+                        $getTargetResourceParameters['IgnoreAuthenticationErrors'] = $false
+                    }
+
+                    It 'Should throw the correct error' {
+                        { Get-TargetResource @getTargetResourceParameters } | Should -Throw 'The user name or password is incorrect.'
+                    }
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is set to $true' {
+                    BeforeAll {
+                        Mock -CommandName Write-Warning
+                    }
+
+                    BeforeEach {
+                        $getTargetResourceParameters['IgnoreAuthenticationErrors'] = $true
+                    }
+
+                    It 'Should return the same values as passed as parameters' {
+                        $result = Get-TargetResource @getTargetResourceParameters
+                        $result.DomainName | Should -Be $mockDomainName
+                        $result.Credential.UserName | Should -Be $mockUserName
+                    }
+
+                    It 'Should output a warning message' {
+                        $null = Get-TargetResource @getTargetResourceParameters
+
+                        Assert-MockCalled -CommandName Write-Warning -Exactly -Times 1 -Scope It
+                    }
+                }
+            }
         }
         #endregion
 
@@ -599,15 +666,84 @@ try
 
                 It 'Should not throw and call the correct mocks' {
                     Invoke-Command -ScriptBlock $script:waitForDomainControllerScriptBlock -ArgumentList @(
-                        'contoso.com' # DomainName
-                        'Europe', # SiteName
-                        $mockDomainUserCredential, # Credential
                         $true # RunOnce
+                        'contoso.com' # DomainName
+                        'Europe' # SiteName
+                        $mockDomainUserCredential # Credential
                     )
 
                     Assert-MockCalled -CommandName Find-DomainController -Exactly -Times 1 -Scope It
                     Assert-MockCalled -CommandName Clear-DnsClientCache -Exactly -Times 1 -Scope It
                     Assert-MockCalled -CommandName Start-Sleep -Exactly -Times 1 -Scope It
+                }
+            }
+
+            Context 'When the Find-DomainController throws an authentication exception' {
+                BeforeAll {
+                    Mock -CommandName Find-DomainController -MockWith {
+                        $exceptionWithInnerException = New-Object -TypeName 'System.Management.Automation.MethodInvocationException' `
+                        -ArgumentList @(
+                            'The user name or password is incorrect.',
+                            (New-Object -TypeName 'System.Security.Authentication.AuthenticationException')
+                        )
+
+                        $newObjectParameters = @{
+                            TypeName     = 'System.Management.Automation.ErrorRecord'
+                            ArgumentList = @(
+                                $exceptionWithInnerException,
+                                'AuthenticationException',
+                                'InvalidOperation',
+                                $null
+                            )
+                        }
+
+                        throw New-Object @newObjectParameters
+                    }
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is not specified' {
+                    It 'Should throw the correct error' {
+                        {
+                            Invoke-Command -ScriptBlock $script:waitForDomainControllerScriptBlock -ArgumentList @(
+                                $true # RunOnce
+                                'contoso.com' # DomainName
+                                'Europe' # SiteName
+                                $mockDomainUserCredential # Credential
+                            )
+                        } | Should -Throw 'The user name or password is incorrect.'
+                    }
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is set to $false' {
+                    It 'Should throw the correct error' {
+                        {
+                            Invoke-Command -ScriptBlock $script:waitForDomainControllerScriptBlock -ArgumentList @(
+                                $true # RunOnce
+                                'contoso.com' # DomainName
+                                'Europe' # SiteName
+                                $mockDomainUserCredential # Credential
+                                $false
+                            )
+                        } | Should -Throw 'The user name or password is incorrect.'
+                    }
+                }
+
+                Context 'When the parameter IgnoreAuthenticationErrors is set to $true' {
+                    BeforeAll {
+                        Mock -CommandName Write-Warning
+                    }
+
+                    It 'Should output a warning message' {
+                        Invoke-Command -ScriptBlock $script:waitForDomainControllerScriptBlock -ArgumentList @(
+                                $true # RunOnce
+                                'contoso.com' # DomainName
+                                'Europe' # SiteName
+                                $mockDomainUserCredential # Credential
+                                $true
+                            )
+
+                            Assert-MockCalled -CommandName Write-Warning -Exactly -Times 1 -Scope It
+                    }
                 }
             }
 
@@ -620,10 +756,10 @@ try
 
                 It 'Should not throw and call the correct mocks' {
                     Invoke-Command -ScriptBlock $script:waitForDomainControllerScriptBlock -ArgumentList @(
-                        'contoso.com' # DomainName
-                        'Europe', # SiteName
-                        $null, # Credential
                         $true # RunOnce
+                        'contoso.com' # DomainName
+                        'Europe' # SiteName
+                        $null # Credential
                     )
 
                     Assert-MockCalled -CommandName Find-DomainController -Exactly -Times 1 -Scope It